add: improved performance by eliminating client components as possible with structure changes
