--- src/app/[exhibitionId]/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-    return null;
-  }
